a lot of great changes

+++ src/main/java/ua/softserve/ita/service/impl/CompanyServiceImpl.java
@@ -4,6 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ua.softserve.ita.dao.*;
 import ua.softserve.ita.dto.CompanyDTO.CompanyPaginationDTO;
+import ua.softserve.ita.exception.ResourceNotFoundException;
 import ua.softserve.ita.model.Company;
 import ua.softserve.ita.model.User;
 import ua.softserve.ita.model.enumtype.Status;
@@ -51,10 +52,8 @@ public class CompanyServiceImpl implements CompanyService {
 
     @Override
     public Optional<Company> save(Company company) {
-
-        User loggedUser = userDao.findById(getLoggedUser().get().getUserID()).get();
+        User loggedUser = userDao.findById(getLoggedUser().get().getUserId()).orElseThrow(() -> new ResourceNotFoundException("User not found"));
         company.setUser(loggedUser);
-
         Optional<Company> com = companyDao.findByName(company.getName());
         Company result = null;
 
@@ -69,15 +68,20 @@ public class CompanyServiceImpl implements CompanyService {
 
     @Override
     public Company update(Company company) {
-        companyDao.update(company);
-        addressDao.update(company.getAddress());
-        contactDao.update(company.getContact());
+        if(company.getUser().getUserId().equals(getLoggedUser().get().getUserId())) {
+            companyDao.update(company);
+            addressDao.update(company.getAddress());
+            contactDao.update(company.getContact());
+        }
         return companyDao.update(company);
     }
 
     @Override
     public void deleteById(Long id) {
-        companyDao.deleteById(id);
+        Company company = companyDao.findById(id).orElseThrow(() -> new ResourceNotFoundException(String.format("Company with id: %d not found", id)));
+        if(company.getUser().getUserId().equals(getLoggedUser().get().getUserId())) {
+            companyDao.deleteById(id);
+        }
     }
 
     @Override
@@ -97,7 +101,7 @@ public class CompanyServiceImpl implements CompanyService {
         if(res.isPresent()) {
             company = res.get();
 
-            if (company.getUser().getUserId().equals(getLoggedUser().get().getUserID())) {
+            if (company.getUser().getUserId().equals(getLoggedUser().get().getUserId())) {
                 company.setStatus(Status.APPROVED);
 
                 User user = company.getUser();
